requirements.txt
aiohttp
boto3
botocore
python-dotenv
requests
